Use a stable ordering for renamed symbols
We should generate the exact same bytecode regardless of the order in which
files are loaded. Relying on Symbol.id for ordering reflects the order in which
symbols were accessed, and that depends on what files are compiled. Instead,
we should use source positions wherever possible, since all the renamable symbols
originate from the currently compiled unit.